Move revoking ca to separate service part
